refactor(collector): separate collection orchestration from height runner
