+++ collector/height_runner.go
@@ -1,86 +1,55 @@
 package collector
 
 import (
-	"fmt"
+	"errors"
 	"time"
 
-	"github.com/dezswap/cosmwasm-etl/pkg/logging"
+	"github.com/dezswap/cosmwasm-etl/collector/repo"
+	"github.com/dezswap/cosmwasm-etl/parser/dex"
 )
 
-type heightCollector interface {
-	LocalHeight() (uint64, error)
-	SourceHeight() (uint64, error)
-	CollectHeight(height uint64) error
+type sourceHeightCollector struct {
+	repo                 repo.Repository
+	source               dex.SourceDataStore
+	chainID              string
+	startHeight          uint64
+	poolSnapshotInterval uint
 }
 
-type heightCollectorConfig struct {
-	StartHeight  uint64
-	UntilHeight  uint64
-	PollInterval time.Duration
+func (c *sourceHeightCollector) LocalHeight() (uint64, error) {
+	localHeight, err := c.repo.GetSyncedHeight(c.chainID)
+	if err == nil {
+		return localHeight, nil
+	}
+	if errors.Is(err, repo.ErrNotFound) || errors.Is(err, repo.ErrUnavailable) {
+		return c.startHeight - 1, nil
+	}
+	return 0, err
 }
 
-// collectHeights runs the common contiguous-height collection loop.
-// Implementations own source reads and persistence for a single height, while
-// the runner handles local/source progress, until-height bounds, and polling.
-func collectHeights(collector heightCollector, config heightCollectorConfig, logger logging.Logger) error {
-	startHeight := config.StartHeight
-	if config.UntilHeight > 0 && config.UntilHeight < startHeight {
-		return fmt.Errorf("invalid height range: start_height=%d until_height=%d", startHeight, config.UntilHeight)
+func (c *sourceHeightCollector) SourceHeight() (uint64, error) {
+	return c.source.GetSourceSyncedHeight()
+}
+
+func (c *sourceHeightCollector) CollectHeight(height uint64) error {
+	txs, err := c.source.GetSourceTxs(height)
+	if err != nil {
+		return err
 	}
 
-	pollInterval := config.PollInterval
+	blockTime := time.Time{}
+	if len(txs) > 0 {
+		blockTime = txs[0].Timestamp
+	}
 
-	for {
-		localHeight, err := collector.LocalHeight()
+	savePoolSnapshot := c.poolSnapshotInterval > 0 && height%uint64(c.poolSnapshotInterval) == 0
+	poolInfos := []dex.PoolInfo{}
+	if savePoolSnapshot {
+		poolInfos, err = c.source.GetPoolInfos(height)
 		if err != nil {
 			return err
 		}
-
-		srcHeight, err := collector.SourceHeight()
-		if err != nil {
-			return err
-		}
-
-		targetHeight := boundedTargetHeight(srcHeight, config.UntilHeight)
-		if localHeight >= targetHeight {
-			if reachedUntilHeight(localHeight, config.UntilHeight) {
-				logger.Infof("collector reached until height %d", config.UntilHeight)
-				return nil
-			}
-			logger.Infof("no new collector source height: local=%d source=%d", localHeight, srcHeight)
-			time.Sleep(pollInterval)
-			continue
-		}
-
-		nextHeight := localHeight + 1
-		if nextHeight < startHeight {
-			nextHeight = startHeight
-		}
-		if nextHeight > targetHeight {
-			logger.Infof("no collectible height yet: local=%d source=%d start=%d", localHeight, srcHeight, startHeight)
-			time.Sleep(pollInterval)
-			continue
-		}
-
-		for height := nextHeight; height <= targetHeight; height++ {
-			if height < startHeight {
-				continue
-			}
-			if err := collector.CollectHeight(height); err != nil {
-				return err
-			}
-			logger.Infof("collected source height %d", height)
-		}
 	}
-}
 
-func boundedTargetHeight(sourceHeight, untilHeight uint64) uint64 {
-	if untilHeight > 0 && untilHeight < sourceHeight {
-		return untilHeight
-	}
-	return sourceHeight
-}
-
-func reachedUntilHeight(localHeight, untilHeight uint64) bool {
-	return untilHeight > 0 && localHeight >= untilHeight
+	return c.repo.SaveHeight(c.chainID, height, blockTime, txs, poolInfos, savePoolSnapshot)
 }
